requirements.txt
matplotlib
pmdarima
prophet
psycopg2-binary
sqlalchemy
python-dotenv
pandas
wheel
scikit-learn
nbconvert
seaborn
traitlets
scipy
plotly
watchdog
mpld3
streamlit==1.18.0 # streamlit==1.11.1
Pillow
streamlit-autorefresh
jupyterflow
https://files.pythonhosted.org/packages/ab/88/dc597fdadd74b1a2fe7d4e4fb56f44a2b444a01b76ec35f31c7eee28f2d0/ray-2.1.0-cp310-cp310-manylinux2014_x86_64.whl #ray[default]==2.1.0
https://github.com/agapebondservant/ml-metrics-accelerator/zipball/main
lz4
pyyaml==5.3.1
prometheus-client
protobuf # protobuf==3.20.0 # protobuf==3.19.5
fastapi
#modin[ray]
uvicorn[standard]
gunicorn
boto3
tensorflow
rsocket
# asyncio
python-snappy
mlflow==2.4.1
filelock==3.9.0
pyarrow<7.0.0
hickle
find-libpython
python-multipart
evidently
pysqlite3-binary
kfp
download_and_extract
altair==4.0
https://files.pythonhosted.org/packages/26/f5/5f0aaf4cbd0017258619a0bfe117e1263035bd501a480eead0799c140025/onnxruntime-1.16.0-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
skl2onnx
tf2onnx
imbalanced-learn
pika
acryl-datahub[datahub-rest]
